Shift chunk_endorsement_tracker into separate module with tracker_v1
