+++ chain/client/src/stateless_validation/mod.rs
@@ -1,4 +1,4 @@
-pub mod chunk_endorsement_tracker;
+pub mod chunk_endorsement;
 pub mod chunk_validator;
 pub mod partial_witness;
 mod shadow_validate;
